backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
agency-swarm>=0.2.0
python-telegram-bot==13.7
python-dotenv
